Signal not existing repository with gRPC code `NotFound`
We implement the check in the `helper.GetRepoPath` method. We could
check the output of each command instead and save us the extra check,
but this would complicate our output parsing, and because of the
decentralized nature of git commands not all of them have the same
output format (for example, the error output when a repository doesn't
exists differs for `git-for-each-ref` and `git-diff`).