Makefile: add ".NOTPARALLEL:"
Several of the dependency lists say "clean-build ${BUILD_DIR}/_build",
which is intended to clean the build dir, then recreate it.  Of course, if
the user has `make` set to build dependencies in parallel, then that won't
work.

Now, I'm not a fan of putting "clean" as a dependency of a constructive
step.  It's sloppy.  But hey, it's not my project.