FindLicense Server Implementation
As part of gitlab-org/gitaly#1026, the server has been implemented. The
current implementation redirects the request to the ruby server as thats
the way its done at the CE/EE codebase too[1], and thus quickest right
now. The same gem is used on Gitaly, which means we should try to keep
those in sync.

Vendoring of gitaly-proto is locked to a branch at the moment, as the
version required has not been released yet.

CI will not run for this commit, as the gitaly-proto gem still needs to
be updated in a commit following this one.
[ci skip]

[1]: https://gitlab.com/gitlab-org/gitlab-ce/blob/216d21c5496fb7e7666c3c3e7ebbdff1e8fd0493/app/models/repository.rb#L589-601