Use danger-gitlab to auto review merge requests
Given failures to check simple things in the past, like if a changelog
was present, danger-gitlab is introduced to automatically perform these
tasks.

A warning was added when a changelog entry is skipped, and failures are
due to missing MR descriptions or gitaly-proto versions being incorrect.

To force diff highlighting to work, `.gitattributes` was added.