Introduce small feature flag package
Feature flags on the Gitaly server weren't possible before and it was
suggested in https://gitlab.com/gitlab-org/gitaly/issues/1170 that
Gitaly could do this through gRPC metadata.

This gives control on wether of not to execute new code paths to Rails,
where the server needs to set a header in the form of
`gitaly-feature-<flag_name>` to "true".