Only load Wiki formatted data upon request
Previously the formatted data for each Wiki page requested would be generated
unnecessarily. This would be expensive especially if the Wiki rendering
required shelling out to another process, as reStructuredText files does. Now,
we only load the formatted data and rely on the separate RPC to request the
formatted data.

Closes https://gitlab.com/gitlab-org/gitlab-ce/issues/45860