Reset sanitize requirements in lockfile
Honestly not quite sure how this number got bumped, but it seems that
the minimum version didn't get bumped but the correct version got used
by bundler.

The report was out of date, as the scanner got removed. However, GitLab
tried to find a pipeline with scanning enabled, and did so from before
it got removed.

So long story short, this was never broken, and while I fixed it I broke
something. I _think_ it has to do with the `--no-deployment` flag.