Bump Code Quality CI/CD Job
Since Rubocop was introduced pipelines succeeded with warnings, as
Rubocop was newer than the quality checks that ran.

This commit replaces the sample with the one listed in the docs:
https://docs.gitlab.com/ee/ci/examples/code_quality.html

The biggest change is the fact that this does now run a version of these
checks based on CI_SERVER_VERSION. This might break our pipeline, or
create new offences at a moments notice. Given `allow_failure` is still
set to true, this could be alright and saves maintainence time.