Prevent stale packed-refs file when Gitaly is running on top of NFS
Users of the Linux NFS client v4.0 (and even v4.1) may encounter stale file
handles or stale data if git gc is running for a long time. A patch for this
kernel bug is here: https://marc.info/?l=linux-nfs&m=153816500525563&w=2

This workaround calls stat() on the packed-refs file to force the kernel to do
a directory entry validation on the inode.

Closes https://gitlab.com/gitlab-org/gitaly/issues/1339