Make git hooks self healing
If the git hooks aren't symlinked properly, GitLab authentication and
authorization is skipped completely. So on most Git operations, if these
hooks are missing this commit will try to repair them when triggering an
upload-pack, or receive-pack git action.

This change also allows the Gitaly team to change the hooks in following
releases to any new ones. And allowing for absorbing the gitlab-shell
hooks into this project.

Previously attempted in:
https://gitlab.com/gitlab-org/gitaly/commit/87198d32837, which ran into
race conditions. This approach does not

Part of https://gitlab.com/gitlab-org/gitaly/issues/1226