Introduce Create, and Delete ObjectPool RPC
For object pools the creation and removal will be dictated by
gitlab-rails, which will do so by making requesting a RPC.

Create will clone a repository to the object pools path, and remove the
origin remote. Also, the local refs are deleted. This does not remove
the objects, which can be reused when linking the repository in a later
RPC reqeust.

The deletion is done by removing the git repository. This might
leave empty parent directories behind. Solving this requires locking or
other mechanisms to prevent race conditions.

Closes https://gitlab.com/gitlab-org/gitaly/issues/1344