Improve Linking and Unlink object pools RPC
GitLab-Rails performs operations when Linking and Unlinking. This makes sure
we perform less RPCs from Rails, and keeps the responsibilities straight. To do
so, config options needed to be set. An RPC was available to set it
through Ruby, but proved quite cumbersome to use as the rubyserver now
needed to go into the `objectpool` package.

Fixes: https://gitlab.com/gitlab-org/gitaly/issues/1420

As a matter of applying the boyscout rule, this commit also sets the
gc.auto option for object pools to 0. This means that Git will never
perform GC without being told to do so.

Fixes: https://gitlab.com/gitlab-org/gitaly/issues/1423