Allow repositories to be reduplicated
When a fork of a GitLab project changes visibility level from public to
anything else, the objects shouldn't be shared anymore. This means that
the repository needs to be reduplicated in terms of git objects.

By executing a repack with all objects, git will make all objects local
again to the repository. Afterwards the alternates can be unset.

Given the diff was small, this commit includes the protobuf upgrade.

Part of: https://gitlab.com/gitlab-org/gitaly/issues/1348