Fix 503 errors when Git outputs warnings to stderr
Git can emit warnings when a command is run, such as when an ambiguous
ref is used or a deprecated feature is used (e.g. `/info/grafts`).
Gitaly will parse these warnings as OIDs and attempt to deference them,
resulting in an error, "2:TypeError: The given OID is too long".

To fix this, we provide an optional parameter to control the
concatenation of the stderr stream to the resulting output. In cases
where we want to see the full message (e.g. when we log an error
message), we include the stderr output. Otherwise, by default for
Repository commands, we disable this flag.

Closes https://gitlab.com/gitlab-org/gitaly/issues/1329