Add glProjectName to logs
This enables the Gitaly logs to have the full human-readable GitLab
project path, which is helpful to have for log aggregators when hashed
storage is in use.

glRepository provides an identifier in the form of `project-<id>`, but
this field is blank when a project is created.

Part of https://gitlab.com/gitlab-org/gitaly/issues/1394