Create Praefect command
By introducting a new command, the pass through proxy gets an execution
path. To allow for manual testing, but also for, probably more
important; getting started with integrating into the GitLab
architecture.

For logging we default to logrus, but these aren't structured yet. Which
should be improved later.

Config is added too, by parsing a TOML file. The path to this file is
passed as first argument to the praefect binary. The output is JSON by
default and not yet configurable.

OS signals are handled too, and allow for a Shutdown of the server.

The basic observability was added too, but on the proxy layer this will
not yield too much, other than tracing.