Stop symlinking hooks on repository creation
In an earlier MR[1] hooks are executed not through the symlink, but
leveraging the `-c` flag on the `git` binary. This works well, but when
`#run_git` was executed in Gitaly-Ruby the hooks weren't executed the
new way.

Luckily this was covered by tests, and the symlinking strategy remained
employed. This commit fixes the bug where the hooks weren't executed and
now allows the removal of the code that set the hooks.

Part of: #1226

[1]: !886