Define Praefect Alpha
In the Gitaly team OKRs there's a goal of shipping an alpha stage HA
solution. However, until this commit it was undefined what that would
look like. To clearify this is now written down in the design doc.

The scope is limited, by design. This will allow us to focus our
efforts.