Remove another test that exercises gogit feature flag
As mentioned in https://gitlab.com/gitlab-org/gitaly/merge_requests/899,
we ran an experiment with go-git some time ago. The outcome was not to
use it for now.

The test asserted that we get the same results whether the feature flag
is on or off. Because the feature flag is a no-op now, this test had
become pointless.