Update Gitaly-Proto with protobuf go compiler 1.2
The protobuf for this branch was compiled with a newer version, v1.2.
This had some changes on the the struct fields. A few got added, which
broke tests that performed deep equal tests. For example an
XXX_sizecache field was added to the messages, which sometimes is set
but is not set when settting up a test.

By depending on equals the XXX_sizecache fields were compared, where the
response would set them. By using proto.Equal() as suggested:
https://groups.google.com/d/msg/protobuf/N-elvFu4dFM/6fGxs1v1AgAJ
this problem is fixed.

Struct literal assignments were updated so tests would run again too.

Related to: https://gitlab.com/gitlab-org/gitaly-proto/merge_requests/266