Fix undefined logger panicing
An oversight made the situation possible where the logger had to log a
message while the logger wasn't defined yet. This happened in a case
where no config file was passed for example.

By defining a the default logger this panic doesn't occur anymore. The
configuration options for the logger aren't set yet, so the output is
different than what users might expect, but it's better than a panic.