Ensure parent directory exists before renaming
By debugging why the pipeline failed on a [CE MR][ce-mr], it appeared
that the behaviour of the RenameNamespace RPC was inconsistent with
other RPCs that require a parent directory. This commit fixes that
behaviour so the CE MR can move along afterwards.

Futher, this commit includes minor changes to the namespace service to
clean it up a bit. Not strictly required, but personnally it always
bugged me a bit that RPCs called others RPCs. This service was the
first, or second RPC that I built on this project, so I felt it was nice
to clean up the mess past me left behind.

The helper package now gets two more convenience methods to remove the
need for `helper.ErrInternal(fmt.Errorf("string %v", err))`.

[ce-mr]: https://gitlab.com/gitlab-org/gitlab-ce/merge_requests/25469