Update labeling rules for GitalyBot
Today I noticed that the labeling as applied previously aren't correct
for the throughput measurements GitLab does:
https://about.gitlab.com/handbook/engineering/management/throughput/#implementation

This change updates the labeling rules so from now on this will be
correctly applied. In the case of Gitaly performance is a feature and
not backstage as the point of Gitaly is improving performance.

En passant it fixes a small issue when no changelog entry is present.