Use 3-way merge for squashing commits
A high number of squashes fail because some lines in the target branch
changed in another commit. However, this failure doesn't happen if you
attempt to squash a branch and merge it in this way:

1. git checkout my-branch
2. git rebase -i HEAD~N (where N is some number of commits)
3. git checkout master
4. git merge my-branch

The fourth step does a 3-way merge. To emulate this behavior, we add
`--3way` to the parameters when we attempt to apply the patch onto the
target branch.

Part of https://gitlab.com/gitlab-org/gitlab-ee/issues/10661