git gc invocation: don't write commit-graph with core.commitGraph=true
This amends code just added in 52ec70ee ("GarbageCollect RPC writes
commit graph and enables via config", 2019-04-24).

This isn't needed, and will result in "gc" dying in obscure
circumstances on git versions before v2.22.0 (not released yet) if the
existing graph were to be corrupted. See
git-vcs/git@43d3561805 ("commit-graph write: don't die if the existing
graph is corrupt", 2019-03-25).

This isn't needed, and "git gc" as of v2.22.0 pretty does the
equivalent of core.commitGraph=false now when writing out the graph,
because we don't want a screwy existing graph to ruin our day when we
write the new one.

Also, having written all this explanation I find that this change
isn't technically needed because 52ec70ee misspelled the config key as
"commitGrap" instead of "commitGrap*h*", but let's remove this line
anyway.