Fix socket path not working
The Listen call used a disk path with the prefix `unix:` present. So the
file would be written to a location that would be relative to the
current working directory and had to start with `unix:`. By cleaning the
prefix it would use the user intended path.

Another minor improvement; when the server receives a signal and
succesfully shuts down it no longer exits with a non zero exit code.