Make git work even harder for us
Generally, when we want to "clean" (i.e. remove artifacts
from) the repo, we only want to remove artifacts that
have been explicitly listed in the .gitignore files. This
modification generalizes this pattern so that we can focus
on maintaining .gitignore files across the repo, rather than
have to remember to also updating this makefile.