Allow explicit missing ref error in tests
In https://gitlab.com/gitlab-org/git/commit/014ade748420b074a06dbb7f5fb974b5e6184f4
the behaviour of upload-pack changed to be explicit about missing refs.
In the Gitaly test the assertion was made against the output of the
command being empty. This is no longer the case, so now an assertion is
made against the output.

This itself might break in the furture, but given the test is about hidden
refs it's probably better to have this test break somewhere in the
future than remove the assertion or weaken it.