Use testhelper.CreateTag to have a Git ident
Locally I can't run some tests as Git would like to know the identity of
the user creating new objects. In my case these object creations get the
following message:

```
testhelper.go:184:
        *** Please tell me who you are.

        Run

          git config --global user.email "you@example.com"
          git config --global user.name "Your Name"

        to set your account's default identity.
        Omit --global to set the identity only in this repository.

        fatal: empty ident name (for <zegerjan@callisto.localdomain>) not allowed
```

Now, this leverages system packages on the Git side to determine the
identname. In our case I think the best way to move forward is just to
use the `testhelper.CreateTag()`, which sets the ident for us.