Changed files used for sparse checkout
To speed up the rebase operation, a sparse checkout could be used.
Meaning that the worktree used to perform the rebase, will not include
lots of files. For rebases Git doesn't care about what files it has
checked out in the sparse checkout, and will, as needed checkout
additional files.

Sadly, git doesn't support an empty sparse checkout. Which is why we
select the changed blobs. These are the ones that are likely needed by
Git.