Log all diskcache state with gRPC metadata
This adds an additional log message that will improve visibility of
whether the diskcache is serving stale data or not. The "freshness" of
the cache is determined by the repo's cache state file. The contents of
this file will now be logged whenever it is updated.

Additionally, the path of the cached response on disk will be logged
each time the diskcache is used. This can be used to help determine if
stale data is being served after the cache is expected to be invalid.
Each time the state file changes, we expect a new request file path.