Strip invalid characters in signatures on commit
When creating a commit, remove angle brackets (`<>`) and newlines (`\n`)
from the name and email fields for the author and committer. This aligns
the behaviour with Git CLI.

This addresses https://gitlab.com/gitlab-org/gitlab/issues/19684, where
committing fails if the user's display name contains angle brackets.
The decision was to keep supporting these characters in GitLab Rails,
because the display name is a free-form text field and we currently
don't restrict the format in any way.

This has also been reported upstream at
https://github.com/libgit2/libgit2/issues/5342