Add `deadline_type` prometheus label to counters
This adds the `deadline_type` prometheus label to the counters from
`go-grpc-prometheus`. The possible values for `deadline_type` are,
`none`, `unknown`, `limited` and `regular`:

- `none`: is set when there is no deadline on the context.
- `unknown`: the default value, this is used when the `deadline_type`
was not included in the headers but a deadline was set
- `limited`: is set when the deadline was shorter than what is
configured. For example when the deadline would exceed the worker
timeout
- `regular`: is set when the deadline is what was configured for the
call in GitLab's settings