Include stderr in output of worktree operations
If a worktree operation fails (e.g. error: Sparse checkout leaves no
entry on working directory), currently no output is stored since the
error message is written to stderr. The only way to see the error
message is to run `strace` on the gitaly-ruby instance.

With this change, the stderr output will be included, which will make
diagnosing a bug much easier.

Relates to https://gitlab.com/gitlab-org/gitaly/issues/2395