Standardize validation errors to use InvalidArgument
Since Praefect throws an InvalidArgument error when it sees an empty
target repository, Gitaly should throw an InvalidArgument as well to be
consistent. This commit also modifies some of the tests to bypass the
gRPC server for testing validation, so we can bypass Praefect all
together. Praefect as a proxy will not have the same error string as the
gitaly server when an empty repository is detected since praefect will
return right away beore the request reaches Gitaly.