Replication: propagate RenameRepository RPC to Praefect secondaries
The RenameRepository is a mutator operation, however it is not about
changing Git data, but a metadata of the repository. That is why it
is must be treated specially same as RemoveRepository operation.
Besides to make a replication an additional data required - the
parameter of the initial request(new repository name).
Previously Replicator marked all jobs that has same relative path
as Cancelled which is not correct because source and target may be
different for each job entry.

Closes: https://gitlab.com/gitlab-org/gitaly/issues/2172