Enforce diff.noprefix=false for generating Git diffs
If the local config had the Git config `diff.noprefix` set to `true`,
then Git would omit the `a/` and `b/` prefixes from the diff, which
would cause a `diff header regexp mismatch` error.

We now ensure that this config is set to `false` to allow local
configurations to coexist with Gitaly.

Closes https://gitlab.com/gitlab-org/gitaly/issues/2477