Handle ambiguous refs in CommitLanguages
When a tag and branch of the same name exist, e.g. tags/master and
heads/master, git rev-parse will default to use the tag name.
If the tag is a tag object, git-linguist will fail with error
'the requested type does not match the type in the ODB'.

This change checks if git rev-parse has warned of an ambiguous ref
and will explicitly use 'heads/<REFNAME>' to ensure the correct ref
is resolved.