Use millisecond precision for JSON time field
Updates the `time` field in JSON formatted logs to use millisecond
precision. Currently second precision is used, but more granularity
is useful with high transaction volume. This will bring Gitaly in
line with the precision used in the rails JSON logs.

For comparison, Gitaly non-JSON logs list the time with 10-microsecond
precision at the start of the line, followed by a `time` field
that uses second precision:

2019-10-04_18:55:54.73359 time="2019-10-04T18:55:54Z