git: uploadpack: Introduce function to enable filter options
When enabling filters for git-upload-pack(1), we need two pass two
config flags to it. While we currently only have a single location where
we need to do this, we're about to add multiple callsites that will need
to pass the same set of flags. This commit thus refactors them into a
function that can be called in order to avoid duplication.