git: stats: Refactor code to pull out HTTP ref announcement parsing
The git/stats includes code to analyze HTTP traffic, most importantly a
parser for the HTTP info-refs protocol that announces capabilities and
references to a client. While the parser is useful outside of this
package, it's currently part of `goGet()`, which will also perform the
actual request for you.

This commit pulls out the parsing code into its own function that gets
as input an `io.Reader` interface. This makes it reusable in other parts
of the code base. Furthermore, this also refactors the parsing code to
use an explicit state machine.