Praefect: use enum values for job states
Move out from `int` as enum values that need to be stored in DB.
Base type for `JobState` and `ChangeType` changed from
`int` type with `iota + 1` initialization to `string`.
It is hard to use binary arithmetic with SQL storage
and it is hard to determine actual value of such type
in the logs.
Unused errors were removed as well.