smarthttp: upload_pack: Use stats pack for packfile negotiation analysis
The smarthttp service keeps track of all "deepen" actions that are
requested via the upload-pack feature. In order to detect them, we have
a function `scanDeepen()` that parses all pktlines part of the
negotiation process and checks whether they are a "deepen" verb.

As we are going to extend our analysis to include additional features
like "have" and "filter" verbs, we have introduced PackfileNegotiation
statistics into our internal git/stats package. So let's drop the custom
implementation that scans for "deepen" instructions in favor of the more
general stats package.

No functional changes are expected as a result of this refactoring.