gitaly-ssh: Refactor `run()` to use a command argument struct
The `run()` function gained quite some of arguments and is going to be
extended by another one to set up feature flags. This makes invocations
of that function hard to read, especially as three of the current
arguments are strings that may be confused quite easily. So let's
refactor this to make use of a new struct `gitalySshCommand` and
implement `run()` as a method on that struct.