gitaly-ssh: Allow passing feature flags via the environment
It is currently not possible to set up any feature flags when using
gitaly-ssh, which we require to test partial clones in our SSH
upload-pack tests. In order to allow for this, let's implement support
for another environment variable "GITALY_FEATUREFLAGS", which is simply
a comma-separated list of flags that gitaly-ssh should set up for the
requesting client. Thus, setting up a specific feature flag is as simple
as doing:

    GITALY_FEATUREFLAG=upload_pack_filter gitaly-ssh upload-pack ...