docs: Link GitalySSH Readme to the docs directory
Being under the impression that the gitaly-ssh docs still needed to be
written, I set out to do so. Luckily while I was reading the source I
saw the doc existed as a readme file.

The index now links to a document outside of the docs directory, though
I expect our users to be able to handle that.