git: stats: Implement analysis for packfile negotiation
Similar to the reference discovery code, this commit adds a new parser
for the packfile negotiation protocol. It analyses a client stream and
records various information sent by the client like the haves,
capabilities and total number of pktlines sent. This will at a later
stage be used to analyze incoming packfile negotiations and fill in
Prometheus metrics to increase visibility of what features are used by
clients.