ssh: upload_pack: Refactor tests to make use of command struct
The `testClone()` function currently performs both the required setup to
do a clone as well as the cleanup afterwards. While this is handy in
most cases, it requires us to shove all tests we want to do on the
repository into that function, as the repo is otherwise lost upon
return.

Improve the situation by implementing a new struct `cloneCommand` that
collects all parameters required to perform a clone. This struct has two
methods, one of which is `execute()` that will perform the request
generation as well as the clone itself, and one method `test()` that has
the same semantics as the previous `testClone()` function.  Like this,
one may continue to use `cmd.test()` in case checks provided by that
function are sufficient and use `cmd.clone()` if one needs to verify
additional things.