ssh: upload_pack: Refactor `cloneCommand.test()` to always assert success
The `cloneCommand.test()` function returns an error code in case any
error happened, but there is only a single caller that will _not_
immediately call `require.NoError()` on that error code. As such, let's
refactor that caller to use `cloneCommand.execute()` instead and then
convert `cloneCommand.test()` to assert for success itself to remove the
burden for callers. This lifts the burden for all other callers and
makes code easier to read.