Use Rugged::Repository#bare over #new
For a different issue I was reading part of the Rugged code base, and
stumbled upon a comment implying that opening a repository using #bare
was faster. I did not validate this through benchmarks, but given the
commit history and implementation of the method, I don't see a reason
that it would harm us.