linguist: check cmd status before parsing output
When running Git linguist on a repository, Gitaly tries to parse output
which is not intended to be parsed as the exit code might be
non-zero. Gitaly expect success on the command, and JSON output. On an
error, the output is an error message string, not in JSON.

This creates parsing errors, and issues like:
https://gitlab.com/gitlab-org/gitaly/-/issues/2294. The current change
checks the exit status of the command, where the expectation is that
linguist will never output invalid JSON if the exit code is zero.
Other cases will bubble up the error string in the error itself.