grpc-proxy: Let codec satisfy both grpc.Codec and encoding.Codec
The `grpc.Codec` interface has been deprecated upstream with commit
5ba054bf (encoding: Introduce new method for registering and choosing
codecs (#1813), 2018-01-23) in favor of the new `encoding.Codec`
interface. We're still using the old interface throughout Gitaly, and in
fact we cannot currently fully migrate to using the old interface as
there is no way to create `grpc.ServerOption`s for the new interface.

In order to allow the conversion of most of our code to the new
interface, let our own codec type satisfy both the old `grpc.Codec` and
new `encoding.Codec` interface. The only thing that has changed between
both is the `String()` function, which has been renamed to `Name()`. It
is thus trivial to support both at the same time.

As the new `Codec` type collides with the `Codec()` function, rename the
latter to `NewCodec()`.