grpc: Convert to use `grpclog.SetLoggerV2()`
The function `grpclog.SetLogger()` has been deprecated upstream in favor
of a new and more flexible `grpclog.SetLoggerV2()` that can
differentiate the log level. As the standard `log.Logger` interface does
not have the `Error` family of functions implemented, we cannot use it
out of the box as a logger for grpclog. We could implement these
functions ourselves on top of `log.Logger`, but that'd mean we have to
implement three functions. Instead, let's make use of the helper
function `grpclog.NewLoggerV2()`, that simply accepts an `io.Writer`.